Fix method naming, parameter types, add missing annotations

+++ service/switchboardapp/src/main/java/com/switchboard/app/repository/DecoderRepository.java
@@ -3,7 +3,7 @@ package com.switchboard.app.repository;
 import com.switchboard.app.domain.Decoder;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DecoderRepository extends JpaRepository<Decoder, Integer> {
+public interface DecoderRepository extends JpaRepository<Decoder, Long> {
