+++ service/switchboardapp/src/main/java/com/switchboard/app/repository/EncoderRepository.java
@@ -3,7 +3,7 @@ package com.switchboard.app.repository;
 import com.switchboard.app.domain.Encoder;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EncoderRepository extends JpaRepository<Encoder,Integer> {
+public interface EncoderRepository extends JpaRepository<Encoder,Long> {
 
 
 }
